Better login page & Quick Connect error handling (#43)
Related to #18 

Ensures that when trying to auth via Quick Connect, the button to use a
username/password instead is always shown.

Additionally, errors that occur during the quick connect process are
displayed instead of just logging in the background.

This PR does not change the Quick Connect behavior or code, it's just UI
improvements for when an error occurs with Quick Connect.